requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
gitpython==3.1.43
python-multipart==0.0.9
PyGithub==2.2.0
python-dotenv==1.0.0 